test/cuda/requirements.txt
tensorflow-gpu==2.4.1
torch